Fix fish
After the flurry of last night’s commits, this morning nothing with fish was working.
It didn’t know about brew, nvim, or code. I found this https://stackoverflow.com/questions/66724016/my-fish-is-blind-fish-does-not-recognise-any-commands-after-setting-it-as-defa stack overflow.
Adding `fish_add_path /opt/homebrew/bin` made it recognize the commands but then it would forget the next time I opened a fish session.
Nick helped me determine that fish was no longer linked so we stowed it, then added the homebrew path to the config.fish.
Now it will always know where that lives. It was probably stored somewhere in the myriad of files that I deleted and wanted git to stop tracking.